reorg solar fixtures and mark test types
